requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
blobfile==1.1.1
bpe==1.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
DALL-E==0.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
hypothesis==6.8.1
idna==2.10
importlib-metadata==3.7.3
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.2.0
mypy==0.812
mypy-extensions==0.4.3
nltk==3.5
numpy==1.20.1
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
py==1.10.0
pycryptodomex==3.10.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
regex==2021.3.17
requests==2.25.1
sip==6.0.3
six==1.15.0
sortedcontainers==2.3.0
toml==0.10.2
toolz==0.11.1
torch==1.8.0
torchvision==0.9.0
tqdm==4.59.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
xmltodict==0.12.0
zipp==3.4.1
